adding verbose build flags

+++ source/json.c
@@ -13,7 +13,7 @@ void insert_spaces(uint8_t level) {
 
 void printf_object(struct json_object_t object, uint8_t level) {
   printf("{\n");
-  for (int i=0; i<object.size; ++i) {
+  for (size_t i=0; i<object.size; ++i) {
     insert_spaces(level+INC);
     printf("%s: ", object.keys[i]->literal);
     printf_json_value(*(object.values[i]), level);
@@ -24,7 +24,7 @@ void printf_object(struct json_object_t object, uint8_t level) {
 
 void printf_array(struct json_array_t array, uint8_t level) {
   printf("[\n");
-  for (int i=0; i<array.size; ++i) {
+  for (size_t i=0; i<array.size; ++i) {
     insert_spaces(level+INC);
     printf_json_value(*(array.elements[i]), level);
   }
@@ -99,19 +99,18 @@ void free_json_value(struct json_value_t *value) {
     default:
       exit(67);
   }
-  free(value); 
+  free(value);
 }
 
 void free_json_array(struct json_array_t *array) {
-  for (int i=0; i<array->size; ++i) {
+  for (size_t i=0; i<array->size; ++i) {
     free_json_value(array->elements[i]);
   }
   free(array);
 }
 
-
 void free_json_object(struct json_object_t *object) {
-  for (int i=0; i<object->size; ++i) {
+  for (size_t i=0; i<object->size; ++i) {
     free_json_string(object->keys[i]);
     free_json_value(object->values[i]);
   }
